requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
astroid==3.2.2
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
configparser==7.0.0
dill==0.3.8
discord.py==2.3.2
frozenlist==1.4.1
google-api-core==2.19.0
google-api-python-client==2.48.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
Google-Images-Search==1.4.7
googleapis-common-protos==1.63.0
httplib2==0.22.0
idna==3.7
isort==5.13.2
mccabe==0.7.0
mistralai==0.0.9
multidict==6.0.5
oauthlib==3.2.2
pillow==10.3.0
platformdirs==4.2.2
praw==7.7.1
prawcore==2.4.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyfiglet==0.8.post1
pylint==3.2.2
pyparsing==3.1.2
python-dotenv==1.0.1
python-resize-image==1.1.20
requests==2.32.3
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
termcolor==1.1.0
together==0.2.6
tomlkit==0.12.5
tweepy==4.14.0
update-checker==0.18.0
uritemplate==4.1.1
urllib3==2.2.1
websocket-client==1.8.0
yarl==1.9.4
